Fixed fill_map function

+++ lib/libft/ft_strlen.c
@@ -6,14 +6,14 @@
 /*   By: dvaisman <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/03 13:33:15 by dvaisman          #+#    #+#             */
-/*   Updated: 2022/10/03 13:34:35 by dvaisman         ###   ########.fr       */
+/*   Updated: 2023/05/25 17:05:58 by dvaisman         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-#include "libft.h"
 
-size_t	ft_strlen(const char *str)
+
+int	ft_strlen(const char *str)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (str[i])
